pyproject.toml
[project]
name = "musickit"
version = "0.1.0"
description = "Convert FLAC/MP3 rips to Apple Lossless (ALAC) with clean tags + embedded covers."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "typer>=0.24.2",
  "mutagen>=1.47.0",
  "rich>=14.0.0",
  "pillow>=11.0.0",
  "httpx>=0.28.1",
  "pydantic>=2.10.0",
]

[project.scripts]
musickit = "musickit.cli:app"

[dependency-groups]
dev = [
  "ruff>=0.15.12",
  "mypy>=1.20.2",
  "pyright>=1.1.409",
  "pytest>=9.0.3",
  "pytest-cov>=7.1.0",
  "coverage[toml]>=7.13.5",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "W", "F", "I", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"**/__init__.py" = ["D104"]
"src/**/*.py" = ["D102", "D105", "D107"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git", ".venv", "__pycache__"]

[tool.coverage.run]
branch = true
relative_files = true
source = ["musickit"]

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:"]
precision = 2
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
strict_equality = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["mutagen.*"]
ignore_missing_imports = true

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/.venv"]
pythonVersion = "3.13"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportPrivateUsage = false
reportUnusedFunction = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportUnknownLambdaType = false
reportMissingImports = "warning"
reportMissingModuleSource = false
